+++ test/Util.hs
@@ -1,6 +1,7 @@
 module Util where
 
 import           Data.List
+import           Data.Char
 import           Test.Hspec.Meta
 import qualified System.IO.Silently as S
 
@@ -18,3 +19,12 @@ x `shouldStartWith` y = x `shouldSatisfy` isPrefixOf y
 
 shouldEndWith :: (Eq a, Show a) => [a] -> [a] -> Expectation
 x `shouldEndWith` y = x `shouldSatisfy` isSuffixOf y
+
+-- replace times in summary with zeroes
+normalizeSummary :: [String] -> [String]
+normalizeSummary xs = map f xs
+  where
+    f x | "Finished in " `isPrefixOf` x = map g x
+        | otherwise = x
+    g x | isNumber x = '0'
+        | otherwise  = x